+++ __tests__/setup/pageObjects/hideCursor.js
@@ -1,0 +1,3 @@
+export default async function hideCursor(driver) {
+  await driver.executeScript(() => document.querySelector(':focus').blur());
+}